changing temperature type to int

+++ src/main/java/org/pfaa/chemica/model/Constants.java
@@ -2,7 +2,7 @@ package org.pfaa.chemica.model;
 
 public interface Constants {
 	/* Technically Standard Ambient Temperature and Pressure (SATP) */
-	public final static double STANDARD_TEMPERATURE = 298; // K
+	public final static int STANDARD_TEMPERATURE = 298; // K
 	public final static double STANDARD_PRESSURE = 101; // kPa
 	
 	public final static double R = 0.008314; /* (m^3*kPa) / (K*mol) */
